requirements.txt
anyio==3.7.1
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
dnspython==2.4.0
exceptiongroup==1.1.2
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
memphis-py==1.0.5
nats-py==2.3.1
protobuf==4.23.4
pymongo==4.4.1
referencing==0.29.1
rpds-py==0.8.10
sniffio==1.3.0
tqdm==4.65.0

# Additional
python-dotenv==1.0.0
black==23.7.0
flake8==6.1.0
isort==5.12.0
pandas==2.0.3
pyarrow==12.0.1
ipykernel
plotly
nbformat
gensim
scikit-learn
seaborn
nltk
statsmodels
supabase
psycopg2
dbt-postgres
streamlit

# Current package
-e .